refactor: 在线状态订阅、连接层与消息同步批次优化，补充测试
- 在线状态：订阅改走 WS 2005（含 Dart 侧服务/Provider/测试），
  失败降级处理与重订阅
- 连接层：reader/rpc 重构、hub 事件分发调整、心跳与重连完善
- 消息同步：PullMessageBySeqs 同会话多 SeqRange 覆盖问题修复
  （批次切分保证同会话单区间）
- 消息处理：撤回/通知/发送/查询路径清理，DB 层调整
- 测试：新增在线状态、fixtures（black_list/group_members/user_status）
  与 Dart 测试，更新既有集成测试

注：库代码 cargo check 通过；测试模块存在重构后残留引用
（WS_SUB_USER_ONLINE_STATUS / parse_revoke_tips_from_json 等），
待后续修复。

+++ rust/src/message/operate/delete.rs
@@ -189,9 +189,9 @@ mod tests {
         let m1 = dao.get_by_client_msg_id("conv_1", "m1").await.unwrap().unwrap();
         let m3 = dao.get_by_client_msg_id("conv_1", "m3").await.unwrap().unwrap();
         let m2 = dao.get_by_client_msg_id("conv_1", "m2").await.unwrap().unwrap();
-        assert_eq!(m1.status, msg_status::HAS_DELETED as i32);
-        assert_eq!(m3.status, msg_status::HAS_DELETED as i32);
-        assert_eq!(m2.status, msg_status::SEND_SUCCESS as i32);
+        assert_eq!(m1.status, msg_status::HAS_DELETED);
+        assert_eq!(m3.status, msg_status::HAS_DELETED);
+        assert_eq!(m2.status, msg_status::SEND_SUCCESS);
         // Deleted 事件发布
         match msg_rx.try_recv().unwrap() {
             MessageEvent::Deleted { conversation_id, client_msg_ids } => {
@@ -220,7 +220,7 @@ mod tests {
         assert_eq!(api.delete_count(), 1, "即使 seqs 为空也会调用服务端");
         // 本地仍软删
         let m = dao.get_by_client_msg_id("conv_1", "m_local").await.unwrap().unwrap();
-        assert_eq!(m.status, msg_status::HAS_DELETED as i32);
+        assert_eq!(m.status, msg_status::HAS_DELETED);
     }
 
     #[tokio::test]
